[MIGRATE] change `retargeting_pixel_id ` to big_int

+++ db/migrate/20170324164535_add_retargeting_pixel_id_to_pages_and_articles.rb
@@ -1,6 +1,6 @@
 class AddRetargetingPixelIdToPagesAndArticles < ActiveRecord::Migration[5.0]
   def change
-    add_column :pages, :retargeting_pixel_id, :integer
-    add_column :articles, :retargeting_pixel_id, :integer
+    add_column :pages, :retargeting_pixel_id, :bigint
+    add_column :articles, :retargeting_pixel_id, :bigint
   end
 end
